mobile friendly and styled better
fixed the styling to look better as well as allowed for it to be viewable in moble.

Ryan i had to change some of your javascript so that it can accomidate for mobile.
so instead of it clearing the order tab when you click "x" from the drop down  it now clears it whenever you click a new pizza this way it can work with mobile.